Jiten latest

+++ src/main/java/com/simplevat/rest/transactioncategorybalancecontroller/TransactioncategoryBalancePersistModel.java
@@ -1,6 +1,7 @@
 package com.simplevat.rest.transactioncategorybalancecontroller;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import lombok.Data;
 
@@ -10,5 +11,5 @@ public class TransactioncategoryBalancePersistModel implements Serializable{
 	private Integer transactionCategoryBalanceId;
 	private Integer transactionCategoryId;
 	private String effectiveDate; // dd/MM/yyyy
-	private Double openingBalance;
+	private BigDecimal openingBalance;
 }
